requirements.txt
click
requests
marshmallow
pytz
shortuuid
tabulate
kafka-python
pathlib2
tzlocal
progressbar33
websocket-client